+++ common/src/main/java/net/zelythia/autotools/AutoTools.java
@@ -196,6 +196,7 @@ public class AutoTools {
             }
             inventory.selected = sourceSlot;
 
+            if(!AutoToolsConfig.get().switchBack) swaps.clear();
             return;
         }
 
@@ -458,7 +459,8 @@ public class AutoTools {
                         || !blockState.requiresCorrectToolForDrops()) return;
             }
 
-            for (int i = 0; i < inventory.getContainerSize(); i++) {
+            int containerLimit = AutoToolsConfig.get().hotbarOnly? 9: inventory.getContainerSize();
+            for (int i = 0; i < containerLimit; i++) {
                 Item item = inventory.getItem(i).getItem();
 
                 if (item != Items.AIR) {
@@ -493,7 +495,40 @@ public class AutoTools {
                 }
             }
 
-            if (toolSlot == -1 || DO_NOT_SWAP_UNLESS_ENCH.contains(BuiltInRegistries.BLOCK.getKey(blockState.getBlock())) && miningSpeed.priority == 0) {
+            if(DO_NOT_SWAP_UNLESS_ENCH.contains(BuiltInRegistries.BLOCK.getKey(blockState.getBlock())) && miningSpeed.priority == 0){
+                return;
+            }
+            else if(toolSlot == -1) {
+                // If we haven't found a valid tool, select an item with no durability
+                if(AutoToolsConfig.get().preserveDurability) {
+                    if(inventory.getSelected().getMaxDamage() == 0) return;
+
+
+                    if(AutoToolsConfig.get().switchBack && !swaps.empty()){
+                        int firstItem = swaps.get(0);
+                        if(inventory.getItem(firstItem).getMaxDamage() == 0){
+                            selectItem(client, inventory, firstItem);
+                            return;
+                        }
+                    }
+
+                    // Prefer empty hand over other items in hotbar
+                    for (int i = 0; i < 9; i++) {
+                        if(inventory.getItem(i).isEmpty()){
+                            selectItem(client, inventory, i);
+                            return;
+                        }
+                    }
+
+                    int containerLimit1 = AutoToolsConfig.get().hotbarOnly? 9: inventory.getContainerSize();
+                    for (int i = 0; i < containerLimit1; i++) {
+                        if(inventory.getItem(i).getMaxDamage() == 0){
+                            selectItem(client, inventory, i);
+                            return;
+                        }
+                    }
+
+                }
             } else {
                 selectItem(client, inventory, toolSlot);
             }
@@ -509,7 +544,8 @@ public class AutoTools {
                 return;
             }
 
-            for (int i = 0; i < inventory.getContainerSize(); i++) {
+            int containerLimit = AutoToolsConfig.get().hotbarOnly? 9: inventory.getContainerSize();
+            for (int i = 0; i < containerLimit; i++) {
                 Item item = inventory.getItem(i).getItem();
 
                 if (item != Items.AIR) {