+++ common/src/main/java/net/zelythia/autotools/config/AutoToolsConfig.java
@@ -132,10 +132,16 @@ public class AutoToolsConfig extends PartitioningSerializer.GlobalData {
         public boolean keepSlot = false;
 
         @ConfigEntry.Gui.Tooltip
+        public boolean hotbarOnly = false;
+
+        @ConfigEntry.Gui.Tooltip
         public boolean preferHotbarTool = true;
 
         @ConfigEntry.Gui.Tooltip
         public boolean preferLowDurability = false;
+
+        @ConfigEntry.Gui.Tooltip
+        public boolean preserveDurability = true;
 
         @ConfigEntry.Gui.Tooltip
         public boolean alwaysPreferFortune = false;